Move the local invocation routine to local.rs
